requirements.txt
pdfplumber
openai==1.55.3
langchain==0.1.0
langchain-community==0.0.20
langchain-openai
python-dotenv
pyyaml
pandas